feat : query 저장 api 추가 - query 저장 가능 - 들고 와서 , 기준으로 split해 모델에서 사용
