+++ models/__init__.py
@@ -1,4 +1,5 @@
 from models.admin import Admin
+from models.query import Query
 from models.video import Video,VideoFile
 
-__all__ = [Video,Admin,VideoFile]
+__all__ = [Video,Admin,VideoFile,Query]
