+++ models/admin.py
@@ -41,24 +41,3 @@ class AdminData(BaseModel):
         }
 
 
-class Query(Document):
-    text: str
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "text": "woman man",
-            }
-        }    
-
-
-class TextOutput(BaseModel):
-    words: list[str]
-    word_count: int
-
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "words": ["woman","man"],
-                "word_count" : 2,
-            }
-        }
